caitsith/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caitsith"
version = "0.1.0"
description = "Funciones tipo Excel para pandas DataFrames con app interactiva opcional."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pandas",
  "numpy",
]

[project.optional-dependencies]
app = [
  "streamlit>=1.44",
  "openpyxl>=3.1",
  "pyyaml>=6.0",
]
notebook = [
  "ipython>=8.0",
  "ipywidgets>=8.0",
]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
